% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{MetaT}
\alias{MetaT}
\title{Gaze and board data from a SMI RED 500 eyetracker at 500 Hz.
Data obtained from Meta-T. Excludes identifiers and most "features" typically collected, but otherwise unaltered.}
\format{\describe{
  \item{ts}{game timestamp}
  \item{event_type}{event type of "GAME_EVENT", "EYE_SAMP", etc.}
  \item{game_number}{the current game number for the row}
  \item{episode_number}{the current episode number for the row}
  \item{level}{the current level for the row}
  \item{score}{the current score for the row}
  \item{lines_cleared}{the lines_cleared for the row}
  \item{completed}{was the game completed, only prints for last row of the game}
  \item{game_duration}{game duration, only prints for the last row of the game}
  \item{avg_ep_duration}{average episode duration, only prints for the last row of the game}
  \item{zoid_sequence}{zoid sequence, the sequence of zoid over the course of a game, only prints for the last row of the game}
  \item{evt_id}{first part of a three-part event-identfication. Determines type of event for "GAME_EVENT" events}
  \item{evt_data1}{clarification for the evt_id column}
  \item{evt_data2}{clarification for the evt_data1 column}
  \item{curr_zoid}{the current zoid for the row}
  \item{next_zoid}{the current zoid in the "next box" for the current row}
  \item{board_rep}{board representation the the current time, only prints when the representation changes}
  \item{zoid_rep}{zoid representation the the current time, only prints when the representation changes}
  \item{smi_ts}{sample time}
  \item{smi_eyes}{blink information}
  \item{smi_samp_x_l}{horizontal gaze location of left eye}
  \item{smi_samp_x_r}{horizontal gaze location of right eye}
  \item{smi_samp_y_l}{vertical gaze location of left eye}
  \item{smi_samp_y_r}{vertical gaze location of right eye}
  \item{smi_diam_x_l}{width of left pupil}
  \item{smi_diam_x_r}{width of right pupil}
  \item{smi_diam_y_l}{height of left pupil}
  \item{smi_diam_y_r}{height of right pupil}
  \item{smi_eye_x_l}{horizontal offset of the left eye relative to the center of the screen}
  \item{smi_eye_x_r}{horizontal offset of the right eye relative to the center of the screen}
  \item{smi_eye_y_l}{vertical offset of the left eye relative to the center of the screen}
  \item{smi_eye_y_r}{vertical offset of the right eye relative to the center of the screen}
  \item{smi_eye_z_l}{distance of the left eye to the center of the screen}
  \item{smi_eye_z_r}{distance of the right eye to the center of the screen}
}}
\usage{
data(MetaT)
}
\description{
\itemize{
 \item Samplerate was 500hz.
 \item Screen resolution was 1680 x 1050 (px).
 \item Screen size was 473.76 x 296.1 (mm).
}
}
\examples{
data(MetaT)
MetaT

}
\keyword{datasets}
