% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{MetaTWO}
\alias{MetaTWO}
\title{Board and gameplay data obtained from Meta-TWO, unaltered, meaning that there are no meaningful headers.
However, V3, V4, V5, V6 have been removed as they identify the subject}
\format{\describe{
  \item{V1}{game timestamp}
  \item{V2}{event type of "GAME_EVENT", "EYE_SAMP", etc.}
  \item{V7}{the current game number for the row}
  \item{V8}{the current episode number for the row}
  \item{V9}{the current level for the row}
  \item{V10}{the current score for the row}
  \item{V11}{the lines_cleared for the row}
  \item{V12}{was the game completed, only prints for last row of the game}
  \item{V13}{game duration, only prints for the last row of the game}
  \item{V14}{average episode duration, only prints for the last row of the game}
  \item{V15}{zoid sequence, the sequence of zoid over the course of a game, only prints for the last row of the game}
  \item{V16}{first part of a three-part event-identfication. Determines type of event for "GAME_EVENT" events}
  \item{V17}{clarification for the evt_id column}
  \item{V18}{clarification for the evt_data1 column}
  \item{V19}{the current zoid for the row}
  \item{V20}{the current zoid in the "next box" for the current row}
  \item{V21}{ARE, referred to as "entry delay" }
  \item{V22}{DAS, a.k.a. "delayed auto-shift" count.}
  \item{V23}{softdrop count}
  \item{V24}{board representation the the current time, only prints when the representation changes}
  \item{V25}{zoid representation the the current time, only prints when the representation changes}


}}
\usage{
data(MetaTWO)
}
\description{
Board and gameplay data obtained from Meta-TWO, unaltered, meaning that there are no meaningful headers.
However, V3, V4, V5, V6 have been removed as they identify the subject
}
\examples{
data(MetaTWO)
MetaTWO

}
\keyword{datasets}
