% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playEpGame.R
\name{playEpGame}
\alias{playEpGame}
\title{Generate either a widget or video to view zoid behavior over the course of a game.}
\usage{
playEpGame(g, game, episode.col, game.col, board_rep.col, zoid_rep.col,
  time.col = NULL, video = F, suffix = NULL, .interval = 0.1,
  justTime = F)
}
\arguments{
\item{g}{A data object or string location. Used to form an object of class \code{"data.table"} that holds zoid_rep and board_rep information.}

\item{game}{A numeric. Used to select which game you are interested in.}

\item{episode.col}{A string. Used to designate the name of the column from the object \code{g} that contains Episode Numbers.}

\item{game.col}{A string. Used to designate the name of the column from the object \code{g} that contains Game Numbers.}

\item{board_rep.col}{A string. Used to designate the name of the column from the object \code{g} that contains Board Representations.}

\item{zoid_rep.col}{A string. Used to designate the name of the column from the object \code{g} that contains Zoid Representations.}

\item{time.col}{A string. Used to designate the name of the column from the object \code{g} that contains Time stamp information. Setting this alters the behavior of the manipulate version of the function, uses time instead of "frame" as the second slider.}

\item{video}{Baoolean (default \code{FALSE}). Used to determine whether the output will be a widget or a video.}

\item{suffix}{A string used if \code{video} is \code{TRUE}. The suffix is added to the end of the video file name for clarity. Defaults to \code{NULL}, and if \code{NULL} the suffix will be the current datestamp.}

\item{.interval}{A numeric set to determine the duration of a frame for video playback (when \code{video} is \code{TRUE}). Defaults to 0.1. When \code{time.col} is provided, and \code{video} is False, sets the increment the slider will move at.}

\item{justTime}{A boolean. Used in conjunction with \code{time.col} such that when \code{time.col} is provided changes manipulate widget functionality to only use one slider for time.}
}
\value{
Either a Shiny widget or Video depending on the designation of \code{video} (and whether the \code{"animation"} package loads).
}
\description{
Generate either a widget or video to view zoid behavior over the course of a game.
}
\examples{

rda <- fread("data.txt") ## load in the file, or use sample data
data(MetaTWO)
rda <- MetaTWO

#option one, do not include "time.col". This has episodes as the top slider, and frames as the bottom slider.
playEpGame(g=rda, game = 3,game.col="V7",episode.col="V8",board_rep.col = "V24",zoid_rep.col="V25")

#option two, do include "time.col", justTime default to F. This has episodes as the top slider, and time in the episode as the bottom slider. .interval defaults to 0.1
playEpGame(g=rda, game = 3,game.col="V7",episode.col="V8",board_rep.col = "V24",zoid_rep.col="V25", time.col ="V1")

#option three, do include "time.col", set justTime to T. This has only one slider which is time.  .interval defaults to 0.1
playEpGame(g=rda, game = 3,game.col="V7",episode.col="V8",board_rep.col = "V24",zoid_rep.col="V25", time.col ="V1", justTime = T, .interval = .1)

## The following results in a video. However, it requires the Animation package (and will download it if it is not installed).
##      Further, requires that you externally install FFmpeg (not an R package) as that is the software animation requires to create a video
##      using the function install.ffmpeg() will provide steps on how to do so.  Note that creating a video takes a long time
##      about as long as the game would've taken in realtime. Also note that the interval refers to how long each frame should stay on, which means that
##      you're not actually getting the actual time between each frame.
playEpGame(g=rda, game = 3,game.col="V7",episode.col="V8",board_rep.col = "V24",zoid_rep.col="V25", time.col ="V1", video=T, suffix= "003", .interval = 0.1)

# Additional Example with Built-in Data
data(MetaT) ## loads in example Meta-T complete file, unaltered
playEpGame(g=MetaT, game = 2, game.col = "game_number", episode.col = "episode_number",
           board_rep.col = "board_rep", zoid_rep.col = "zoid_rep")

data(MetaTWO) ## loads in example Meta-TWO game data file, unaltered
playEpGame(g=MetaTWO, game = 2, game.col = "V7", episode.col = "V8",
           board_rep.col = "V24", zoid_rep.col = "V25")

}
